Remove remaining internal namespace leaks from the public include section
