+++ include/kettle/circuit_loggers/classical_register_circuit_logger.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "kettle/circuit/classical_register.hpp"
-#include "kettle_internal/common/clone_ptr.hpp"
+#include "kettle/common/clone_ptr.hpp"
 
 namespace ket
 {
@@ -11,7 +11,7 @@ class ClassicalRegisterCircuitLogger
 public:
     void add_classical_register(ClassicalRegister cregister)
     {
-        cregister_ = ket::internal::ClonePtr<ClassicalRegister> {std::move(cregister)};
+        cregister_ = ket::ClonePtr<ClassicalRegister> {std::move(cregister)};
     }
 
     [[nodiscard]]
@@ -27,7 +27,7 @@ public:
 private:
     // there is no default constructor for the ClassicalRegsiter (it wouldn't make sense), and we
     // only find out how many bits are needed after the first simulation; hence why we use a pointer
-    ket::internal::ClonePtr<ClassicalRegister> cregister_ {nullptr};
+    ket::ClonePtr<ClassicalRegister> cregister_ {nullptr};
 };
 
 }  // namespace ket
