+++ include/kettle/common/tolerance.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace ket
+{
+
+constexpr inline auto NORMALIZATION_TOLERANCE = double {1.0e-6};
+constexpr inline auto COMPLEX_ALMOST_EQ_TOLERANCE_SQ = double {1.0e-6};
+constexpr inline auto MATRIX_2X2_SQRT_TOLERANCE = double {1.0e-6};
+
+}  // namespace ket
