+++ include/kettle/gates/multiplicity_controlled_u_gate.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "kettle/circuit/circuit.hpp"
-#include "kettle/common/mathtools.hpp"
+#include "kettle/common/tolerance.hpp"
 #include "kettle/common/matrix2x2.hpp"
 #include "kettle/common/utils.hpp"
 
@@ -22,7 +22,7 @@ void apply_multiplicity_controlled_u_gate(
     const Matrix2X2& unitary,
     std::size_t target_index,
     const Container& control_indices,
-    double matrix_sqrt_tolerance = ket::internal::MATRIX_2X2_SQRT_TOLERANCE
+    double matrix_sqrt_tolerance = ket::MATRIX_2X2_SQRT_TOLERANCE
 );
 
 }  // namespace ket
