+++ include/kettle/io/write_tangelo_file.hpp
@@ -3,10 +3,7 @@
 #include <cstddef>
 #include <filesystem>
 #include <iostream>
-#include <string>
 
-#include "kettle/common/matrix2x2.hpp"
-#include "kettle/gates/primitive_gate.hpp"
 #include "kettle/circuit/circuit.hpp"
 
 /*
@@ -20,29 +17,6 @@ CNOT      target : [4]   control : [2]
 RZ        target : [5]   parameter : 12.533816585267923
 ```
 */
-
-namespace ket::internal
-{
-
-auto format_double_(double x) -> std::string;
-
-auto format_matrix2x2_(const ket::Matrix2X2& mat) -> std::string;
-
-auto format_one_target_gate_(const ket::GateInfo& info) -> std::string;
-
-auto format_one_control_one_target_gate_(const ket::GateInfo& info) -> std::string;
-
-auto format_one_target_one_angle_gate_(const ket::GateInfo& info) -> std::string;
-
-auto format_one_control_one_target_one_angle_gate_(const ket::GateInfo& info) -> std::string;
-
-auto format_m_gate_(const ket::GateInfo& info) -> std::string;
-
-auto format_u_gate_(const ket::GateInfo& info, const ket::Matrix2X2& mat) -> std::string;
-
-auto format_cu_gate_(const ket::GateInfo& info, const ket::Matrix2X2& mat) -> std::string;
-
-}  // namespace ket::internal
 
 
 namespace ket
