+++ source/kettle_internal/simulation/gate_pair_generator.hpp
@@ -4,7 +4,7 @@
 #include <cstddef>
 #include <tuple>
 
-#include "kettle/common/mathtools.hpp"
+#include "kettle_internal/common/mathtools_internal.hpp"
 
 namespace ket::internal
 {
