+++ test/source/gates/gate_test.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 
-#include "kettle_internal/common/clone_ptr.hpp"
+#include "kettle/common/clone_ptr.hpp"
 #include "kettle/common/matrix2x2.hpp"
 #include "kettle/gates/common_u_gates.hpp"
 #include "kettle/gates/primitive_gate.hpp"
@@ -93,7 +93,7 @@ TEST_CASE("Test create and unpack gate functions")
     SECTION("U gate")
     {
         const auto target_index = std::size_t {0};
-        const auto unitary_ptr = ket::internal::ClonePtr<ket::Matrix2X2> {ket::x_gate()};
+        const auto unitary_ptr = ket::ClonePtr<ket::Matrix2X2> {ket::x_gate()};
 
         const auto gate_info = cre::create_u_gate(target_index, unitary_ptr);
         const auto [unpacked_target_index, unpacked_unitary_ptr] = cre::unpack_u_gate(gate_info);
@@ -107,7 +107,7 @@ TEST_CASE("Test create and unpack gate functions")
         const auto control_index = std::size_t {0};
         const auto target_index = std::size_t {1};
 
-        const auto unitary_ptr = ket::internal::ClonePtr<ket::Matrix2X2> {ket::x_gate()};
+        const auto unitary_ptr = ket::ClonePtr<ket::Matrix2X2> {ket::x_gate()};
 
         const auto gate_info = cre::create_cu_gate(control_index, target_index, unitary_ptr);
         const auto [u_control_index, u_target_index, u_unitary_ptr] = cre::unpack_cu_gate(gate_info);
